+++ Editor/Editors/YarnProjectImporterEditor.cs
@@ -348,6 +348,25 @@ namespace Yarn.Unity.Editor
             };
             yarnInternalControls.Add(addLocalisationButton);
 
+#if USE_ADDRESSABLES
+            var updateAddressableAssetsButton = new Button();
+            updateAddressableAssetsButton.text = "Update Asset Addresses";
+            updateAddressableAssetsButton.clicked += () =>
+            {
+                YarnProjectUtility.UpdateAssetAddresses(yarnProjectImporter);
+            };
+            yarnInternalControls.Add(updateAddressableAssetsButton);
+
+            bool ShouldShowAssetsButton() => yarnProjectImporter.ImportData.localizations.Any(l => l.assetsFolder != null) && AddressableAssetSettingsDefaultObject.SettingsExists && useAddressableAssetsProperty.boolValue;
+
+            SetElementVisible(updateAddressableAssetsButton, ShouldShowAssetsButton());
+
+            useAddressableAssetsField.RegisterCallback<ChangeEvent<bool>>(evt => 
+            {
+                SetElementVisible(updateAddressableAssetsButton, ShouldShowAssetsButton());
+            });
+#endif
+
 #if USE_UNITY_LOCALIZATION
 
             localisationControls.Add(useUnityLocalisationSystemField);
@@ -362,8 +381,6 @@ namespace Yarn.Unity.Editor
                 SetElementVisible(unityControls, useUnityLocalisationSystemProperty.boolValue);
                 SetElementVisible(yarnInternalControls, !useUnityLocalisationSystemProperty.boolValue);
             });
-
-            
 #endif
 
             var cantGenerateUnityStringTableMessage = new IMGUIContainer(() =>
@@ -486,279 +503,7 @@ namespace Yarn.Unity.Editor
         {
             return base.HasModified() || AnyModifications;
         }
-
-        public override void OnInspectorGUI()
-        {
-            DrawDefaultInspector();
-            ApplyRevertGUI();
-#warning Deprecated code
-            return;
-            serializedObject.Update();
-
-            YarnProjectImporter yarnProjectImporter = serializedObject.targetObject as YarnProjectImporter;
-
-            EditorGUILayout.Space();
-
-            if (sourceScriptsProperty.arraySize == 0)
-            {
-                EditorGUILayout.HelpBox("This Yarn Project has no content. Add Yarn Scripts to it.", MessageType.Warning);
-            }
-            EditorGUILayout.PropertyField(sourceScriptsProperty, true);
-
-            EditorGUILayout.Space();
-
-            bool hasCompileError = compileErrorsProperty.arraySize > 0;
-
-            if (hasCompileError)
-            {
-                foreach (SerializedProperty compileError in compileErrorsProperty) {
-                    EditorGUILayout.HelpBox(compileError.stringValue, MessageType.Error);
-                }
-            }
-
-            serializedDeclarationsList.DrawLayout();
-
-            // The 'Convert Implicit Declarations' feature has been
-            // temporarily removed in v2.0.0-beta5.
-
-#if false
-            // If any of the serialized declarations are implicit, add a
-            // button that lets you generate explicit declarations for them
-            var anyImplicitDeclarations = false;
-            foreach (SerializedProperty declProp in serializedDeclarationsProperty) {
-                anyImplicitDeclarations |= declProp.FindPropertyRelative("isImplicit").boolValue;
-            }
-            
-            if (hasCompileError == false && anyImplicitDeclarations) {
-                if (GUILayout.Button("Convert Implicit Declarations")) {
-                    // add explicit variable declarations to the file
-                    YarnProjectUtility.ConvertImplicitVariableDeclarationsToExplicit(yarnProjectImporter);
-
-                    // Return here becuase this method call will cause the
-                    // YarnProgram contents to change, which confuses the
-                    // SerializedObject when we're in the middle of a GUI
-                    // draw call. So, stop here, and let Unity re-draw the
-                    // Inspector (which it will do on the next editor tick
-                    // because the item we're inspecting got re-imported.)
-                    return;
-                }
-            }
-#endif
-
-            EditorGUILayout.PropertyField(defaultLanguageProperty, new GUIContent("Base Language"));
-
-            CurrentProjectDefaultLanguageProperty = defaultLanguageProperty;
-
-#if USE_UNITY_LOCALIZATION
-            EditorGUILayout.PropertyField(useUnityLocalisationSystemProperty, new GUIContent("Use Unity's Built-in Localisation System"));
-
-            // if we are using the unity localisation system we need a field to add in the string table
-            // and we also disable the in-built localisation system while we are at it for its unnecessary
-            if (useUnityLocalisationSystemProperty.boolValue)
-            {
-                EditorGUI.indentLevel += 1;
-                EditorGUILayout.PropertyField(unityLocalisationTableCollectionProperty, new GUIContent("String Table Collection"));
-                EditorGUI.indentLevel -= 1;
-                EditorGUILayout.Space();
-            }
-            else
-            {
-                EditorGUILayout.PropertyField(languagesToSourceAssetsProperty, new GUIContent("Localisations"));
-            }
-#else
-            EditorGUILayout.PropertyField(languagesToSourceAssetsProperty, new GUIContent("Localisations"));
-#endif
-
-
-            // Determines whether or not we draw the GUI for the internal Yarn
-            // localisation system.
-            bool showInternalLocalizationGUI;
-
-#if USE_UNITY_LOCALIZATION
-            // If Unity Localization is available, then we draw the internal
-            // localization system's UI if we've chosen not to use Unity's.
-            showInternalLocalizationGUI = !useUnityLocalisationSystemProperty.boolValue;
-#else
-            // If Unity Localization is not available, we always draw the
-            // internal localization system UI.
-            showInternalLocalizationGUI = true;
-#endif
-
-            CurrentProjectDefaultLanguageProperty = null;
-
-            // Ask the project importer if it can generate a strings table.
-            // This involves querying several assets, which means various
-            // exceptions might get thrown, which we'll catch and log (if
-            // we're in debug mode).
-            bool canGenerateStringsTable;
-
-            try
-            {
-                canGenerateStringsTable = yarnProjectImporter.CanGenerateStringsTable;
-            }
-            catch (System.Exception e)
-            {
-#if YARNSPINNER_DEBUG
-                Debug.LogWarning($"Encountered in error when checking to see if Yarn Project Importer could generate a strings table: {e}", this);
-#else
-                // Ignore the 'variable e is unused' warning
-                var _ = e;
-#endif
-                canGenerateStringsTable = false;
-            }
-
-            if (showInternalLocalizationGUI)
-            {
-                // The following controls only do something useful if all of the
-                // lines in the project have tags, which means the project can
-                // generate a string table.
-                using (new EditorGUI.DisabledScope(canGenerateStringsTable == false))
-                {
-#if USE_ADDRESSABLES
-
-                    // If the addressable assets package is available, show a
-                    // checkbox for using it.
-                    var hasAnySourceAssetFolders = yarnProjectImporter.ImportData.localizations.Any(l => l.assetsFolder != null);
-                    if (hasAnySourceAssetFolders == false)
-                    {
-                        // Disable this checkbox if there are no assets
-                        // available.
-                        using (new EditorGUI.DisabledScope(true))
-                        {
-                            EditorGUILayout.Toggle(useAddressableAssetsProperty.displayName, false);
-                        }
-                    }
-                    else
-                    {
-                        EditorGUILayout.PropertyField(useAddressableAssetsProperty);
-
-                        // Show a warning if we've requested addressables but
-                        // haven't set it up.
-                        if (useAddressableAssetsProperty.boolValue && AddressableAssetSettingsDefaultObject.SettingsExists == false)
-                        {
-                            EditorGUILayout.HelpBox("Please set up Addressable Assets in this project.", MessageType.Warning);
-                        }
-                    }
-
-                    // Add a button for updating asset addresses, if any asset
-                    // source folders exist
-                    if (useAddressableAssetsProperty.boolValue && AddressableAssetSettingsDefaultObject.SettingsExists)
-                    {
-                        using (new EditorGUI.DisabledScope(hasAnySourceAssetFolders == false))
-                        {
-                            if (GUILayout.Button($"Update Asset Addresses"))
-                            {
-                                YarnProjectUtility.UpdateAssetAddresses(yarnProjectImporter);
-                            }
-                        }
-                    }
-#endif
-                }
-            }
-
-            EditorGUILayout.Space();
-
-#if YARN_USE_LEGACY_ACTIONMANAGER
-
-            EditorGUILayout.LabelField("Commands and Functions", EditorStyles.boldLabel);
-
-            var searchAllAssembliesLabel = new GUIContent("Search All Assemblies", "Search all assembly definitions for commands and functions, as well as code that's not in a folder with an assembly definition");
-            EditorGUILayout.PropertyField(searchAllAssembliesProperty, searchAllAssembliesLabel);
-
-            if (searchAllAssembliesProperty.boolValue == false)
-            {
-                EditorGUI.indentLevel += 1;
-                EditorGUILayout.PropertyField(assembliesToSearchProperty);
-                EditorGUI.indentLevel -= 1;
-            }
-
-            EditorGUILayout.PropertyField(predeterminedFunctionsProperty, true);
-#endif
-
-            if (showInternalLocalizationGUI)
-            {
-
-                using (new EditorGUI.DisabledGroupScope(canGenerateStringsTable == false))
-                {
-                    if (GUILayout.Button("Export Strings and Metadata as CSV"))
-                    {
-                        var currentPath = AssetDatabase.GetAssetPath(serializedObject.targetObject);
-                        var currentFileName = Path.GetFileNameWithoutExtension(currentPath);
-                        var currentDirectory = Path.GetDirectoryName(currentPath);
-
-                        var destinationPath = EditorUtility.SaveFilePanel("Export Strings CSV", currentDirectory, $"{currentFileName}.csv", "csv");
-
-                        if (string.IsNullOrEmpty(destinationPath) == false)
-                        {
-                            // Generate the file on disk
-                            YarnProjectUtility.WriteStringsFile(destinationPath, yarnProjectImporter);
-
-                            // Also generate the metadata file.
-                            var destinationDirectory = Path.GetDirectoryName(destinationPath);
-                            var destinationFileName = Path.GetFileNameWithoutExtension(destinationPath);
-
-                            var metadataDestinationPath = Path.Combine(destinationDirectory, $"{destinationFileName}-metadata.csv");
-                            YarnProjectUtility.WriteMetadataFile(metadataDestinationPath, yarnProjectImporter);
-
-                            // destinationPath may have been inside our Assets
-                            // directory, so refresh the asset database
-                            AssetDatabase.Refresh();
-                        }
-                    }
-                    if (yarnProjectImporter.ImportData.localizations.Count > 0)
-                    {
-                        if (GUILayout.Button("Update Existing Strings Files"))
-                        {
-                            YarnProjectUtility.UpdateLocalizationCSVs(yarnProjectImporter);
-                        }
-                    }
-                }
-            }
-
-            // Does this project's source scripts list contain any actual
-            // assets? (It can have a count of >0 and still have no assets
-            // when, for example, you've just clicked the + button but
-            // haven't dragged an asset in yet.)
-            var hasAnyTextAssets = yarnProjectImporter.ImportData.yarnFiles.Where(s => s != null).Count() > 0;
-
-            // Disable this button if 1. all lines already have tags or 2.
-            // no actual source files exist
-            using (new EditorGUI.DisabledScope(canGenerateStringsTable == true || hasAnyTextAssets == false))
-            {
-                if (GUILayout.Button("Add Line Tags to Scripts"))
-                {
-                    YarnProjectUtility.AddLineTagsToFilesInYarnProject(yarnProjectImporter);
-                }
-            }
-
-            var hadChanges = serializedObject.ApplyModifiedProperties();
-
-#if UNITY_2018
-            // Unity 2018's ApplyRevertGUI is buggy, and doesn't
-            // automatically detect changes to the importer's
-            // serializedObject. This means that we'd need to track the
-            // state of the importer, and don't have a way to present a
-            // Revert button. 
-            //
-            // Rather than offer a broken experience, on Unity 2018 we
-            // immediately reimport the changes. This is slow (we're
-            // serializing and writing the asset to disk on every property
-            // change!) but ensures that the writes are done.
-            if (hadChanges)
-            {
-                // Manually perform the same tasks as the 'Apply' button
-                // would
-                ApplyAndImport();
-            }
-#endif
-
-#if UNITY_2019_1_OR_NEWER
-            // On Unity 2019 and newer, we can use an ApplyRevertGUI that
-            // works identically to the built-in importer inspectors.
-            ApplyRevertGUI();
-#endif
-        }
-
+        
         public VisualElement CreateUpgradeUI(YarnProjectImporter importer) {
             var ui = new VisualElement();